vectorforge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vectorforge"
version = "0.1.0"
description = "High-performance backtesting engine with hybrid vectorized/event-driven architecture"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "FIREKit Team" }
]
keywords = [
    "backtesting",
    "trading",
    "quantitative-finance",
    "algorithmic-trading",
    "jax",
    "vectorized"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.26.0",
    "pandas>=2.1.0",
    "polars>=0.20.0",
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "numba>=0.58.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "pyarrow>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
gpu = [
    "jax[cuda12_pip]>=0.4.20",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/ZhiruiFeng/FIREKit"
Documentation = "https://github.com/ZhiruiFeng/FIREKit/docs"
Repository = "https://github.com/ZhiruiFeng/FIREKit"
Issues = "https://github.com/ZhiruiFeng/FIREKit/issues"

[tool.hatch.build.targets.wheel]
packages = ["vectorforge"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["vectorforge"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["vectorforge"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
